Updated openstack/openstack Project: openstack/cinder  d1b5a113256bbf5dc246bc781c2dbc1b6ab8e4d3
Fix status_code of "list versions" API

The status code of "list versions" API is 300 because the API method
is decorated with MULTIPLE_CHOICES[1] and the unit test[2] expects the
code. In addition, the corresponding Tempest test[3] also verifies 300
as the code. So it is not necessary to write the other status codes on
the api-ref.

[1]: http://git.openstack.org/cgit/openstack/cinder/tree/cinder/api/versions.py?id=a4a5fac7564503a09751db98d2a4f8be3766577a#n123
[2]: http://git.openstack.org/cgit/openstack/cinder/tree/cinder/tests/unit/api/test_versions.py?id=a4a5fac7564503a09751db98d2a4f8be3766577a#n66
[3]: Id5791665d8435fac7552733b0c1126cee4359d00

Change-Id: Id28e26fca0adbdcfb0d905f52ca697a74db3be75