Updated openstack/openstack Project: openstack/horizon  fb191ef812235089631915c4ead322cd786166e2
Implement "FILTER_DATA_FIRST" in angular views

This patch implements the FILTER_DATA_FIRST setting in angular views
specifically for the resource-table.
As a first approach this was implemented in NG admin > images.

To implement in other views, the following must be done:

- Add the FILTER_DATA_FIRST setting to the
  REST_API_ADDITIONAL_SETTINGS array in the local_settings.py file:
  REST_API_ADDITIONAL_SETTINGS = ['FILTER_DATA_FIRST']

- Enable the FILTER_DATA_FIRST setting:
  FILTER_DATA_FIRST = { 'admin.images': true }

P.S for this to work each panel must implement the check for the
setting accordingly.

Implements: blueprint admin-views-filter-first
(https://blueprints.launchpad.net/horizon/+spec/admin-views-filter-first)

Change-Id: Ieeedffbc621291251d50e7d8f254d3ce4def896c