Updated openstack/openstack Project: openstack/neutron  bc711b1513143aa160c8fd965f64b3f959febbea
Decouple tests from default quota size option values

At this point, if we decide to bump a quota for a resource, we would
need to change the test. It suggests that those tests are not doing the
right thing (they catch any change, not just a regression).

The right thing for the tests to validate is not a particular default
value for a resource quota, but the fact that resources indeed get
default values applied, irrespective of particular values that we tend
to choose at this point in time.

Change-Id: I84afd9e2b9fed8f6eae1743a34bb8064739123a6