Updated openstack/openstack Project: openstack/tripleo-image-elements  17f835b487efd966fef349e9600edbb9e3672fb3
Add overcloud-secure element

This element will export all the block device configuration
needed for security images. As the layout to expose is complex,
it won't fit on the environment vars that can be provided on tripleo-common.
So is better to create a new element that exposes all this configuration,
and include it on the security hardened image.
It will also uninstall some packages that need to be removed for
security reasons.

Change-Id: Id6ece1c734d4cbf5adb857f0e627f59543be44ae
Needed-By: I541055fe81900b91e2bf131f1e95ce08c94f2554