Updated openstack/openstack Project: openstack/tripleo-image-elements  7e5e4dc029198d46705c2e8a4f3f3a06899ab450
51-hosts fails if given lots of changes

The issue is how awk is used to update hosts files.  When
os-apply-config produces sufficiently large amounts of lines
to be added (or ensure in) hosts files, awk will error out.
To work around it, instead use sed, and reconstruct the
host file(s) to ensure the entries between the comment delimeters
of "# HEAT_HOSTS_START" and "# HEAT_HOSTS_END" are swapped with
the new entries.

Also get rid of blank lines produced by os-apply-config

Partial-Bug: #1674732
Change-Id: Ibe0a9f6ec10d55750e3b0e16301236141f988d69