Updated openstack/openstack Project: openstack/nova  2a41eb8786fa1654bbecb50c2a15fdd96fdc4767
Get instance availability_zone without hitting the api db

This would be an upcall from the cell to the api db, which we don't
want. We store the availability_zone on the instance now, and even
use it for pre-scheduled instances, so just use that instead of
the extra lookup.

Related to blueprint cells-aware-api

Change-Id: I73c3b10e52ab4cfda9dacc0c0ba92d1fcb60bcc9