Updated openstack/openstack Project: openstack-dev/devstack  c36b08248e0215dfad078eea64d12e311ed62ec5
Create /opt/stack and make it home directory

single-machine.rst and index.rst

Before this, one had to create /opt/stack
manually and chown it to the stack user.
Now it is created when the user is created.

This is the same way the multi-node guide
handles it. A stack group is created too.

Change-Id: I5363d81c8fb38796f565cc6ebf6ab2dee2673989
Closes-Bug: #1673787
Closes-Bug: #1671409