Updated openstack/openstack Project: openstack/nova  6f881e8b8463c66b49e6361a112c408695caf67f
Revert "libvirt: Pass instance to connect_volume and ..."

This reverts commit f0153fa4c87528e1dcbda8cba55d34ab7aec4e7a.

This broke a whole set of FS-based volume drivers since it
didn't change the method signatures in those drivers to take
the new instance argument.

I was working on adding them but then just decided this
isn't even used right now so it's faster to revert.

When this is added back in, I want to see the
LibvirtBaseVolumeDriver class turned into an ABC and
the connect_volume and disconnect_volume methods turned
into abstractmethods in LibvirtBaseVolumeDriver so this
kind of thing does not happen again.

Change-Id: If1ee6d6e4004750f80424ec54d7dabc564ec9e06
Closes-Bug: #1675584