Updated openstack/openstack Project: openstack/cinder  ce775446ce98b00aec847f4faef76d0200612c3d
Update db from drivers in default groups methods

The following default generic volume group methods in
volume/manager.py are not using returned model_update
from drivers. As a result database is not properly updated
for drivers that are using fields such as provider_id to
save info on volumes/snapshots.

  _create_group_snapshot_generic()
  _create_group_from_src_generic()

This patch fixed the problem.

Change-Id: I118175fc5be0f61146d8762a81113578c338037c
Closes-Bug: #1674454