Updated openstack/openstack Project: openstack/openstack-manuals  851d9c0ada483145584a5ff018f6beb83995520d
PDF theme & styling with openstackdocstheme

This commit applies theme and styling to PDF generation
using xelatex with customized styles.
In openstackdocstheme, style is defined in pdftheme.sty,
and conf.py makes use of the file as LaTeX package.
Also, some rst files are modified to add PDF directives
to adjust better UI in PDFs.

Not all documents are our targets in this commit but feel
free to add more here if anyone wants.

- Current applied styles with openstackdocstheme
 : Dedicated title without a blank page
 : Put a logo image into suitable place.
 : No chapter/section number
 : Light gray background for code blocks
 : Font definition is in openstackdocstheme not in conf.py

- rst modification
 : Table wrap adjustment (".. tabularcolumns::" directive)
 : Exclude "Search in this page" section (".. only:: html" directive)

Change-Id: I2782e0a1f0258a6016559b16c8967cc52f48f734
Depends-On: I423675ff79827966e789df70b0f970ac2a10da50
Implements: blueprint build-pdf-from-rst-guides
Co-Authored-By: Ian Y. Choi <ianyrchoi@gmail.com>