Updated openstack/openstack Project: openstack-infra/project-config  03342c287a5405061c7f102ed1a4f55c23301672
networking-ovn: Make OVS master job non-voting.

This job runs against the master branch of OVS.  It's great at finding
problems, which is also why it should probably be non-voting.  We also
have a job that runs against the latest release branch and we will
continue to gate on that one.

Change-Id: I6a5f71e76e7ff18ad163b4665394f27c35437914
Signed-off-by: Russell Bryant <rbryant@redhat.com>