Updated openstack/openstack Project: openstack-infra/project-config  7c053d1f878d52cc9a2e4213d07b282121397bf7
tripleo: add deploy-guide-jobs into tripleo-docs

tripleo-docs is already our repo for documentation, we're going to
re-use it for storing the deployment guide.

Note: post jobs (that publish) will be added later once we figure how to
deal with our branchless doc which is not a draft but our actual stable
doc.

Change-Id: I8eaf14745d864115b4538f289bcd5b82fd416cd6