Updated openstack/openstack Project: openstack/tempest  a70006b3f0d7e7cfb8a9ac0fe489062b0a511b21
Fix multiple create for multiple networks

Change I953fe9cde89d9e8c968889a823911a206b27a4cd extended
the test to count the number of servers created, but in doing so
it stopped selecting the tenant network, which causes the test
to fail when more then one network is available to tenants.

The logic to select the network should not live in the
common server create helper since it depends on configuration;
the alternative is simply to discover the correct tenant network
in the test and pass it to the create method.

closes-bug: #1675059
Change-Id: I3e97a2d911c151b8df941b11935a547b39331a8f