Updated openstack/openstack Project: openstack-infra/tripleo-ci  b8914bab3422e0282a9d94a65b08a5351f9ce97b
Only force http for repos that allow it

It seems that trying to use http for CBS repos causes a lot of
spurious failures due to the forced redirect to https.  Since we
only care about using http for the delorean repos that allow it,
limit the repo sed to just those urls.

Change-Id: I55ebe1d3963055d27c37cd5e9a299dcf929c84ec
Partial-Bug: 1674681