Updated openstack/openstack Project: openstack/oslo.db  6eda25859f271cc3765e13cfb16378959af2e406
Remove deprecated config option sqlite_db

We deprecated config options sqlite_db in July, 2016 [1], and
it shows up in many projects' config files but these projects
don't use it either, so it's confusing to our users. It's time
to remove it, just use config option connection or slave_connection
to connect to the database.

[1] https://review.openstack.org/#/c/338047/

Closes-Bug: #1329086

Change-Id: Id269d921e40edf95eb977b011f1753f633b79d18