Updated openstack/openstack Project: openstack/cinder  217593c8c389add6ef75ea8fe5a14a577f174fdd
Mark the Violin volume drivers as unsupported

The Violin CI has not reported past our CI requirements.

Checking name: Violin Memory V6000 CI
  last seen: 2016-12-31 12:29:05 (81 days, 20:40:44 old)
  last success:  None

Per Cinder's non-compliance policy, this patch marks the drivers
as unsupported and deprecated and it will be removed if the issue
is not corrected by the next cycle.

Change-Id: I423e06cc50f827f36ba31ddc5914ed658fd2c0dd