Updated openstack/openstack Project: openstack-infra/devstack-gate  f9aa9e7dbadd6e630e575ae7c20ff040aad93f8b
Stop after early failures

The problem identified in Idebf137d3a0bb7541060881e60e2725668811918
was more difficult to spot than it needed to be.  The PROJECT clone
was failing in the setup-workspace-new phase due to an invalid entry,
but this did not stop the script.  It continued on to start running
devstack with a half-cloned library list; devstack then complains
about total red-herrings depending on which random things were not
cloned.

This captures the return codes of these early setup scripts and exits
in a sensible fashion.

A exit_handler function is made for a common exit path (future work;
this could probably be an exit-handler function).  The ara generation
is modified to use $WORKSPACE so it works before or after the log
directory is moved.  We also log the PROJECT list after processsing.

Change-Id: I7bde8fc1f2a7db5bd73635aa197377bf5ac614d2