Updated openstack/openstack Project: openstack-infra/devstack-gate  bbf7c84fbbeffc2722f8aae194cc5aaafe62d61a
Fix DEVSTACK_PROJECT_FROM_GIT for pbr

Change Id2caf9561b361c1d4950856423282535d067e015 modified handling of
DEVSTACK_PROJECT_FROM_GIT to add any projects specified there as
entries to the PROJECT list.

pbr enters this path via the "lib-forward-testing" project-config
macros for the gate-tempest-dsvm-neutron-src-pbr-ubuntu-xenial job.
The result is that the PROJECT list now has an entry for
"openstack/pbr" which does not exist, and thus breaks when the PROJECT
list is cloned.

Prior to this we were just relying on the hard-coded openstack-dev
entry prior to this in the file.

Change-Id: Idebf137d3a0bb7541060881e60e2725668811918