Updated openstack/openstack Project: openstack/manila  ff7e1321e13b81bc127e86bdbe7c8c1ee092a8c6
Move create_manila_accounts to post-config

In before, Manila devstack plugin create the service endpoint in
"extra" stage. This caused issue of other services that depends
on the manila endpoints that are not yet created. This commit
move the endpoint creation to "post-config" stage, which is the
practice of other OpenStack services

Change-Id: I3adfbf78d0792471a29c7ff979aad1418e6137d4