Updated openstack/openstack Project: openstack/barbican  97e1dffeeba75c8d6c7a6d26f844d871662611cc
Compare functional test results as dict, not string

Iteration over a dict is nondeterministic in order, so functional
tests could randomly fail if the expected and actual orders do not
match.  This patch changes the assert compare to be done between
dicts, instead of strings, so the order does not matter.

Change-Id: Ief6a57d6c5222534d647441d8f87c10bf9401930
Closes-bug: 1571804