Updated openstack/openstack Project: openstack-infra/project-config  ab9b9131ecc274872fe5caccb0fbd2b20333e28d
New Project - LOCI

LOCI is a project that builds lightweight OCI compatible images for
OpenStack projects. The images themselves contain only the minimum
required binary packages and code to run an OpenStack service, and are
optimised for incoproration with CI/CD workflows and building with no
external tooling.

The built images contain no configuration or helpers, and are
deployment tool agnostic, helping the orchestration platform take full
ownership of their operation. This is in contrast to existing projects
which are focused on building images en masse in unison with the
projects own associated deployment tooling and configuration
methodology.

Key Features:
 * No custom tooling
 * Non-layered image building approach
 * Native mechanism to build from gerrit patchsets
 * Generic images, independant of deployment tooling

Co-Authored-By: Pete Birley <pete@port.direct>
Change-Id: Icd4a18337e9a12780e4776e9f00cc93718c37e37