Updated openstack/openstack Project: openstack/tripleo-heat-templates  5bcfc8c2a758eced575cf5edf79ab3c19a96c952
Pick dynamically the first node for stack validation

When replacing the controller node with resource id 0,
AllNodesValidation will fail because there is an hardcoded reference
to resource.0. With this commit the id for validation is extracted
dynamically with yaql query, picking the first available.
Thanks to Steven Hardy for pointing to the right direction.

Change-Id: I8f2bfacbc005d948bd31ebd51c3d3df3182d5a3c
Closes-Bug: #1673439