Updated openstack/openstack Project: openstack/nova  afe08af0e0dd380864756dc412f8bfb5d079224f
libvirt: Simplify internal usage of get_instance_disk_info

get_instance_disk_info returns a dict, serialized to a string, but all
internal libvirt callers convert it back to the dict. Considering
existence of _get_instance_disk_info, which returns a dict, but requires
different kinds of parameters, the usage of the method is not
convenient.

This patch does:
- rename _get_instance_disk_info to _get_instance_disk_info_from_xml,
which reflects the method nature;
- declare new _get_instance_disk_info, which returns not serialized
dict;
- make block_device_info parameters of both methods mandatory;
- keep the driver public interface unchanged.

Change-Id: I9c0c6954ba5442414e615225a74df7e42f3d0c98