Updated openstack/openstack Project: openstack/nova  37fa794113179e6d6f52267a3d640651f771cdd1
libvirt: add debug logging in detach_device_with_retry

When we hit failures in this method it can be hard to
sort out from a traceback what actually happened, and
if we failed from the intial detach or if we failed
during the retry loop.

This patch adds some simple debug logging statements
to help give context to what's going on when we need
to debug a failure in this operation.

Change-Id: I392192f139ced077664e54469eead0b0df9ebf4a
Related-Bug: #1673483