Updated openstack/openstack Project: openstack/cinder  fe9a51b441ce9136926f81ed194c206564236a6e
[api-ref] Update snapshot/volume's metadata api-ref

Volume/Snapshot contains two kinds API for it's metadata: POST/PUT

POST is used for creatng or replacing the volume/snapshot's
metadata items that match keys. It doesn't modify items that
are not in the request.

PUT is used for replacing all the volume/snapshot's metadata
with the key-value pairs in the request.

Cinder's api-ref description is wrong. This patch fixed the
description and added the missing POST/PUT request.

Change-Id: I51128d32c9f70ba57d7d4cf0634a16ca566db95f