Updated openstack/openstack Project: openstack/ironic-specs  5ba558aa11f2ae4b1076543249de2e049cad458a
[RFE] Add spec for physical network awareness

https://bugs.launchpad.net/ironic/+bug/1666009

Ironic ports and portgroups correspond to the physical NICs that are
available on baremetal nodes.  Neutron ports are logical ports
attached to tenant and provider networks.  When booting baremetal
nodes that have multiple ports or portgroups, the mapping between
logical and physical ports must account for physical network
connectivity of the system.

This feature aims to make Ironic aware of the 'physical_network'
attribute of Neutron networks and network segments.

Change-Id: Ib4917f6d90d23c6e749f3536d3772e5ee6718971