Updated openstack/openstack Project: openstack/nova  4914abd7d2c33941a67d0aecd9b47db916a2cc48
Pre-add functional tests stub to notification testing

Every time a transformed notification merged, other ones
are getting merge conflict mainly because of [1]. In this
patchset, method calls and method stubs are added in order to
avoid merge conflicts in the future. In this patchset, the
currently in-queue notifications are added.

[1]: https://github.com/openstack/nova/blob/896ee40d27539b3d9de01700176f5ed253d69c83/nova/tests/functional/notification_sample_tests/test_instance.py#L66

Change-Id: I2fed82ca3cd023194b81814ec609c2d2df5e0c03
 Implements: bp versioned-notification-transformation-pike