Updated openstack/openstack Project: openstack/os-refresh-config  ec9d2c32ec3f3cdcfd15a83e1bcde2a3238583ad
Add Constraints support

Adding constraints support to clients is slightly more complex than
services as the clients themselves are listed in upper-constraints.txt
which leads to errors that you can't install a specific version and a
constrained version.

This change adds constraints support by also adding a helper script to
edit the constraints to remove os-refresh-config.

Change-Id: Ia7d3b98fcfa5245e8cf51951e4c2af7ac3541014