Updated openstack/openstack Project: openstack/tempest  4df2338dede31b1df31eb53868ab899c385dcd18
Fix for test_volume_swap

In volume swap procedure, server's BDM is updated after the old volume has
been changed to 'available' and the new volume to 'in-use'. When the old
volume is 'available' and the new volume is 'in-use', it is possible that
the server's BDM has not been updated yet. So it is necessary to wait for
volume swap to make sure that the server's BDM has been updated.

Change-Id: I069a603159ecdec924868f45f12408e55eb235ea