Updated openstack/openstack Project: openstack/nova  357951ca1e41a931a4ecccc7559c6946806ff735
Fix 'server' and 'instance' occurrence in api-ref

some api-ref have sentence like '..for a server instance'
which is inconsistent and confusing as we use server for VM terminology.
I think here 'instance' word is being considered object of server which
is wrong.

We should always mention server only.

part of bp:api-ref-in-rst-pike

Change-Id: I32afe56cfc66b34b76d1f7e1b507d3d5e722e6a1