Updated openstack/openstack Project: openstack/tempest  d11f316eec7c1da8912b10477d515d1affc984ae
Should be more clear for device name opt

We are allowed to config the volume_device_name when
running tests, but not or hypervisors guarantee that
they will respect the user defined device name(KVM),
and tests may fail due to inappropriate device name
set.

We should include this information about this opt
to avoid confusions.

Change-Id: I193940c3ae565dd4803f19a77326006ebdc038cc