Updated openstack/openstack Project: openstack/nova  c844482a098af9ae8f1e4df41b9a4439047ef59d
Ensure instance is in active state after notification test

In versioned notification tests, we should ensure that after
testing a notification in test_instance_action, the instance is in
active state.

Currently, there is no guarantee that instance left in active state
after a notification test. In this patchset, it is ensured, and the
notifications, which violates this rule, are refactored.

Change-Id: Iff0012a4065e61111b69c96e1a4f2074340d9c41
Co-Authored-By: Béla Vancsics <vancsics@inf.u-szeged.hu>
Implements: bp versioned-notification-transformation-pike
Closes-Bug: #1674961