Updated openstack/openstack Project: openstack/nova-specs  558ef292dddc24a24c8dbef184f6c3b7ecc38429
Add spec to use cinder's new attachment API

Cinder has added their new attachment API in Ocata.
This is the spec to describe how Nova can adopt that new API.

The aim is to have a solid base on which we can safely implement support
for Cinder multi-attach.

blueprint cinder-new-attach-apis

Change-Id: I8acc30502ba6e9ff061ca2d80fa73f6da2be7cef