Updated openstack/openstack Project: openstack/barbican  e7b652772afcc43ca6630dbc3fe920d0e336fa88
Fix the creation of endpoints in keystone_data.sh

When using bin/keystone_data.sh to create the barbican data it will
fail creating the endpionts because of invalid syntax for the openstack
endpoint create command.  This CR fixes that, making the public and
internal endpoint commands separate as required by the openstack
command.

Change-Id: I3f6d04888ad816e47e7e5882f13afffbb5b366c0
Closes-Bug: #1676548