Updated openstack/openstack Project: openstack-infra/jenkins-job-builder  d39a3568d3f5654ac5f700b370e2a8094dc6f441
scm: handle "bitbucketweb" browser for mercurial

The Git SCM uses "bitbucketweb".

The Hg SCM also has "bitbucketweb" documented, but uses an un-documented
value instead: "bitbucket".

Make the Hg SCM support "bitbucketweb" so it aligns with the Git SCM and
the Hg documentation.

Continue to handle the old "bitbucket" browser, but add a comment that
this is deprecated.

Change-Id: If4266e35964b2d7c751cce009d195d4c1bcf845a