Updated openstack/openstack Project: openstack/horizon  343cbd19b5e4c17c4c4e86fb6ae86cfde95cb84b
Revert "Add the 'fade' class to Python modals, for animation"

After this commit, "Add Rule" in security group panel is not
setup properly and only a few fields are displayed.
I haven't identified the root cause but it looks better to
revert this commit and tackle the original problem again.

This reverts commit 8be1ec5e706b07b47f13b8c97c70327b25b09552.

Change-Id: I266484da68dd355082144889699dda299ab9f997
Closes-Bug: #1677064
Related-Bug: #1669450