Updated openstack/openstack Project: openstack/neutron  b73be2167880435b6df632e428c23a6fa6af9736
ExtensionTestCase: register correct core plugin

We should register the passed plugin as core only if service_type is
False. Otherwise, the passed plugin is a service plugin, in which case
we should register some other core plugin for the test. The easiest
thing is to use the base db plugin for that matter.

This fixes subproject unit tests because now the core plugin will always
have _get_collection function that is needed by some underlying db api
functions.

Change-Id: Ica421831763eb2f5d25fc068287305b402227b53
Closes-Bug: #1676656